requirements.txt
streamlit
torch
transformers
Pillow
plotly
numpy
matplotlib
requests
